assert in steps
